switch Downloader total download size to long to support >16MB
This was int because it was written arond UrlConnection.getContentLength()
which returns an int.  But that doesn't make sense since this will
definitely handle files large than 16MB.

!647
#1192

+++ app/src/main/java/org/fdroid/fdroid/ProgressListener.java
@@ -19,6 +19,6 @@ import java.net.URL;
  */
 public interface ProgressListener {
 
-    void onProgress(URL sourceUrl, int bytesRead, int totalBytes);
+    void onProgress(URL sourceUrl, long bytesRead, long totalBytes);
 
 }
